test(e2e): consolidate duplicate E2E specs, net -822 lines
Merge overlapping test files to reduce redundant test runs:

- Delete unified-main-window.ts (covered by homepage-features.ts HOME section)
- Merge workflow-lifecycle.ts into workflow.ts (add TC-WF-011)
- Merge settings-persistence.ts into settings.ts (add TC-SET-008/009/010)
- Remove SYNC+DRAG describe blocks from homepage-features.ts (covered by data-sync-window.ts and drag-drop-groups.ts)
- Merge ops-server-status-processes.ts into ops-process-server.ts (add OPS-SS-002/PL-001/PL-002)
- Replace mysql-multi-database.ts + postgres-multi-database.ts with parameterized multi-database.ts
- Trim overlapping smoke steps from data-transfer-journey.ts (Steps 1-3) and schema-diff-journey.ts (Steps 1-2)
- Update wdio.conf.ts suite references and e2e-coverage.md docs

Net: -6 files deleted, +1 file created, 15 files changed, -822 lines
