+++ e2e/specs/homepage-features.ts
@@ -276,343 +276,3 @@ describe('主页 TablePlus 风格 (HOME)', () => {
     await expect(await $('[data-testid="new-connection-dialog"]')).toBeDisplayed();
   });
 });
-
-// ═════════════════════════════════════════════════════════════════════
-// 数据同步窗口 E2E 测试
-// ═════════════════════════════════════════════════════════════════════
-
-describe('数据同步窗口 (SYNC)', () => {
-  let mainWindow: string;
-
-  async function openSyncWindow() {
-    await browser.switchToWindow(mainWindow);
-    try {
-      await browser.keys('Escape');
-    } catch {
-      /* no focused element */
-    }
-    await browser.url('tauri://localhost/window.html?window=data-sync');
-    await browser.pause(1500);
-    await $('[data-testid="data-sync-compare"]').waitForDisplayed({ timeout: 15000 });
-  }
-
-  before(async () => {
-    mainWindow = await browser.getWindowHandle();
-    await ensureConnectionsHome();
-    await browser.pause(500);
-  });
-
-  afterEach(async () => {
-    try {
-      await browser.keys('Escape');
-    } catch {
-      /* no focused element */
-    }
-    await browser.switchToWindow(mainWindow);
-    await ensureConnectionsHome();
-    await browser.pause(300);
-  });
-
-  it('SYNC-001: 应显示源和目标连接选择器', async () => {
-    await openSyncWindow();
-    const body = await $('body').getText();
-    expect(body).toContain(t('sync.source'));
-    expect(body).toContain(t('sync.target'));
-  });
-
-  it('SYNC-002: 应显示标题栏 "数据同步 - DataZen"', async () => {
-    await openSyncWindow();
-    const body = await $('body').getText();
-    expect(body).toContain(t('sync.title'));
-  });
-
-  it('SYNC-003: 应显示 TrafficLights 和 ThemeToggle', async () => {
-    await openSyncWindow();
-    const header = await $('header');
-    await expect(header).toBeDisplayed();
-  });
-
-  it('SYNC-004: 应显示"比较"按钮', async () => {
-    await openSyncWindow();
-    const compareBtn = await $('[data-testid="data-sync-compare"]');
-    await expect(compareBtn).toBeDisplayed();
-  });
-
-  it('SYNC-005: 未选择连接时点击比较应提示', async () => {
-    await openSyncWindow();
-    const compareBtn = await $('[data-testid="data-sync-compare"]');
-    await compareBtn.click();
-    await browser.pause(1000);
-    const err = await $('[data-testid="data-sync-error"]');
-    await expect(err).toBeDisplayed();
-    expect(await err.getText()).toContain(t('sync.selectBoth'));
-  });
-
-  it('SYNC-006: 应显示初始引导文本', async () => {
-    await openSyncWindow();
-    const body = await $('body').getText();
-    expect(body).toContain(t('sync.selectPrompt'));
-  });
-
-  it('SYNC-007: 状态栏应显示数据同步标题', async () => {
-    await openSyncWindow();
-    const status = await $('[data-testid="data-sync-window"]');
-    await expect(status).toBeDisplayed();
-    const body = await $('body').getText();
-    expect(body).toContain(t('sync.title'));
-    expect(body).toContain(t('common.dataSync'));
-  });
-
-  it('SYNC-008: 连接下拉应列出已有连接', async () => {
-    await openSyncWindow();
-    await expect(await $('[data-testid="data-sync-source"]')).toBeDisplayed();
-    await expect(await $('[data-testid="data-sync-target"]')).toBeDisplayed();
-    const connCount = (await invokeBackend<Array<{ id: string }>>('get_connections')).length;
-    expect(connCount).toBeGreaterThan(0);
-  });
-
-  it('SYNC-009: 源和目标选同一连接时应报错', async () => {
-    await openSyncWindow();
-    const bodyBefore = await $('body').getText();
-    if (!bodyBefore.includes('本地 PostgreSQL')) return;
-
-    await browser.execute(() => {
-      const wraps = document.querySelectorAll(
-        '[data-testid="data-sync-source"], [data-testid="data-sync-target"]',
-      );
-      for (const wrap of wraps) {
-        const trigger = wrap.querySelector('button');
-        trigger?.click();
-      }
-    });
-    await browser.pause(300);
-
-    const compareBtn = await $('[data-testid="data-sync-compare"]');
-    await compareBtn.click();
-    await browser.pause(1500);
-    const err = await $('[data-testid="data-sync-error"]');
-    if (await err.isDisplayed()) {
-      expect(await err.getText()).toMatch(/相同|same/i);
-    }
-    try {
-      await browser.keys('Escape');
-    } catch {
-      /* dismiss connection picker overlays */
-    }
-    await browser.pause(200);
-  });
-
-  it('SYNC-010: 离开数据同步视图应回到主工作区', async () => {
-    await openSyncWindow();
-    await ensureConnectionsHome();
-    await expect(await $('[data-testid="workspace-nav-connections"]')).toBeDisplayed();
-    await expect(await $(`input[placeholder="${t('main.searchPlaceholder')}"]`)).toBeDisplayed();
-    const handles = await browser.getWindowHandles();
-    expect(handles.length).toBe(1);
-  });
-});
-
-// ═════════════════════════════════════════════════════════════════════
-// 拖拽连接到不同分组 E2E 测试
-// ═════════════════════════════════════════════════════════════════════
-
-describe('拖拽连接到不同分组 (DRAG)', () => {
-  let mainWindow: string;
-
-  before(async () => {
-    mainWindow = await browser.getWindowHandle();
-    await ensureConnectionsHome();
-    await expandAllGroups();
-    await browser.pause(1000);
-
-    await browser.waitUntil(async () => (await $$('[data-group-header]')).length > 0, {
-      timeout: 5000,
-      timeoutMsg: 'Timed out waiting for groups to load',
-    });
-    await browser.waitUntil(async () => (await $$('[data-conn-item]')).length > 0, {
-      timeout: 5000,
-      timeoutMsg: 'Timed out waiting for connection items',
-    });
-  });
-
-  afterEach(async () => {
-    await browser.switchToWindow(mainWindow);
-    await ensureConnectionsHome();
-    await browser.pause(300);
-  });
-
-  it('DRAG-001: 分组头应有 data-group-header 属性', async () => {
-    const count = await browser.execute(() => {
-      return document.querySelectorAll('[data-group-header]').length;
-    });
-    expect(count).toBeGreaterThan(0);
-  });
-
-  it('DRAG-002: 存在多个分组时拖拽交互可用', async () => {
-    const groups = await $$('[data-group-header]');
-    expect(groups.length).toBeGreaterThanOrEqual(1);
-    if (groups.length >= 2) {
-      const firstGroupText = await groups[0].getText();
-      const secondGroupText = await groups[1].getText();
-      expect(firstGroupText).not.toBe(secondGroupText);
-    }
-  });
-
-  it('DRAG-003: 连接项响应 pointerdown 事件', async () => {
-    const items = await $$('[data-conn-item]');
-    if (items.length === 0) return;
-    await browser.execute(() => {
-      const el = document.querySelector('[data-conn-item]');
-      if (!el) return;
-      const rect = el.getBoundingClientRect();
-      el.dispatchEvent(
-        new PointerEvent('pointerdown', {
-          bubbles: true,
-          cancelable: true,
-          clientX: rect.left + 5,
-          clientY: rect.top + 5,
-          button: 0,
-        }),
-      );
-      window.dispatchEvent(
-        new PointerEvent('pointerup', {
-          bubbles: true,
-          cancelable: true,
-          clientX: rect.left + 5,
-          clientY: rect.top + 5,
-          button: 0,
-        }),
-      );
-    });
-    await browser.pause(300);
-    const searchInput = await $(`input[placeholder="${t('main.searchPlaceholder')}"]`);
-    await expect(searchInput).toBeDisplayed();
-  });
-
-  it('DRAG-004: 通过 pointer 事件模拟拖拽不应崩溃应用', async () => {
-    const groups = await $$('[data-group-header]');
-    if (groups.length < 2) return;
-    const items = await $$('[data-conn-item]');
-    if (items.length === 0) return;
-
-    const result = await browser.execute(() => {
-      const items = document.querySelectorAll('[data-conn-item]');
-      const groupEls = document.querySelectorAll('[data-group-header]');
-      if (items.length === 0 || groupEls.length < 2) return 'skip';
-
-      const src = items[0] as HTMLElement;
-      const target = groupEls[1] as HTMLElement;
-      const srcRect = src.getBoundingClientRect();
-      const tgtRect = target.getBoundingClientRect();
-
-      const sx = srcRect.left + srcRect.width / 2;
-      const sy = srcRect.top + srcRect.height / 2;
-      const tx = tgtRect.left + tgtRect.width / 2;
-      const ty = tgtRect.top + tgtRect.height / 2;
-
-      src.dispatchEvent(
-        new PointerEvent('pointerdown', {
-          bubbles: true,
-          cancelable: true,
-          clientX: sx,
-          clientY: sy,
-          button: 0,
-        }),
-      );
-      for (let i = 1; i <= 5; i++) {
-        window.dispatchEvent(
-          new PointerEvent('pointermove', {
-            bubbles: true,
-            cancelable: true,
-            clientX: sx + ((tx - sx) * i) / 5,
-            clientY: sy + ((ty - sy) * i) / 5,
-          }),
-        );
-      }
-      window.dispatchEvent(
-        new PointerEvent('pointerup', {
-          bubbles: true,
-          cancelable: true,
-          clientX: tx,
-          clientY: ty,
-          button: 0,
-        }),
-      );
-
-      return 'ok';
-    });
-
-    expect(result).toBe('ok');
-    await browser.pause(1500);
-    await expect(await $(`input[placeholder="${t('main.searchPlaceholder')}"]`)).toBeDisplayed();
-  });
-
-  it('DRAG-005: 连接项应标记为可拖拽', async () => {
-    const items = await $$('[data-conn-item]');
-    if (items.length === 0) return;
-
-    const draggable = await browser.execute(() => {
-      const el = document.querySelector('[data-conn-item]') as HTMLElement | null;
-      return el?.getAttribute('draggable') === 'true';
-    });
-    expect(draggable).toBe(true);
-  });
-
-  it('DRAG-006: 拖拽经过连接项时分组区域仍可用', async () => {
-    const items = await $$('[data-conn-item]');
-    const groupEls = await $$('[data-group-header]');
-    if (items.length === 0 || groupEls.length < 2) return;
-
-    await browser.execute(() => {
-      const src = document.querySelector('[data-conn-item]') as HTMLElement;
-      const target = document.querySelectorAll('[data-group-header]')[1] as HTMLElement;
-      if (!src || !target) return;
-
-      const sr = src.getBoundingClientRect();
-      const tr = target.getBoundingClientRect();
-
-      src.dispatchEvent(
-        new PointerEvent('pointerdown', {
-          bubbles: true,
-          cancelable: true,
-          clientX: sr.left + 5,
-          clientY: sr.top + 5,
-          button: 0,
-        }),
-      );
-      window.dispatchEvent(
-        new PointerEvent('pointermove', {
-          bubbles: true,
-          cancelable: true,
-          clientX: sr.left + 20,
-          clientY: sr.top + 20,
-        }),
-      );
-      window.dispatchEvent(
-        new PointerEvent('pointermove', {
-          bubbles: true,
-          cancelable: true,
-          clientX: tr.left + tr.width / 2,
-          clientY: tr.top + tr.height / 2,
-        }),
-      );
-    });
-
-    await browser.pause(300);
-
-    await browser.execute(() => {
-      window.dispatchEvent(
-        new PointerEvent('pointerup', {
-          bubbles: true,
-          cancelable: true,
-          clientX: 0,
-          clientY: 0,
-          button: 0,
-        }),
-      );
-    });
-
-    await expect(await $(`input[placeholder="${t('main.searchPlaceholder')}"]`)).toBeDisplayed();
-    await browser.pause(300);
-  });
-});
