+++ e2e/specs/journeys/schema-diff-journey.ts
@@ -81,46 +81,14 @@ describe('结构对比完整用户旅程 (SD-JOURNEY)', () => {
     await browser.switchToWindow(mainWindow);
   });
 
-  it('Step 1: 通过 URL 打开结构对比窗口', async () => {
-    await openSchemaDiffWindow();
-    const body = await $('body').getText();
-    expect(body).toContain(t('schemaDiff.step.endpoints'));
-    expect(body).toContain(t('schemaDiff.step.plan'));
-    expect(body).toContain(t('schemaDiff.step.deploy'));
-    await captureJourneyStep('sd-journey-01-window-open', 0, true);
-  });
-
-  it('Step 2: 选择源/目标后未选表点下一步应提示必填', async () => {
-    await seedSecondPgConnection(browser);
-    await openSchemaDiffWindow();
-    await selectSchemaDiffEndpoints('本地 PostgreSQL', 'E2E-PG-目标');
-    await captureJourneyStep('sd-journey-02-endpoints-selected', 0, true);
-    await clickSchemaDiffNext();
-    const objectsPanel = await $('[data-testid="schema-diff-objects-panel"]');
-    await objectsPanel.waitForDisplayed({ timeout: 15000 });
-    await browser.waitUntil(
-      async () => (await $$('[data-testid="schema-diff-table-row"]').length) > 0,
-      { timeout: 20000, timeoutMsg: '等待表列表加载' },
-    );
-    const rows = await $$('[data-testid="schema-diff-table-row"]');
-    for (const row of rows) {
-      const checkbox = await row.$('input[type="checkbox"]');
-      if (await checkbox.isSelected()) await checkbox.click();
-    }
-    await clickSchemaDiffNext({ requireEnabled: false });
-    const body = await $('body').getText();
-    expect(body).toContain(t('schemaDiff.tableRequired'));
-    await captureJourneyStep('sd-journey-03-table-required', 0, true);
-  });
-
-  it('Step 3: 选择同步专用连接并填写表名', async () => {
+  it('Step 1: 选择同步专用连接并填写表名', async () => {
     await openSchemaDiffWindow();
     await selectSchemaDiffEndpoints(SRC_NAME, TGT_NAME);
     await setSchemaDiffTables(TABLE);
     await captureJourneyStep('sd-journey-04-table-entered', 0, true);
   });
 
-  it('Step 4: 对比结构应显示差异面板', async () => {
+  it('Step 2: 对比结构应显示差异面板', async () => {
     await clickSchemaDiffCompare();
     const body = await $('body').getText();
     expect(body).toContain(TABLE);
@@ -128,7 +96,7 @@ describe('结构对比完整用户旅程 (SD-JOURNEY)', () => {
     await captureJourneyStep('sd-journey-05-compare-result', 0, true);
   });
 
-  it('Step 5~7: 生成计划→审阅→部署并验证 extra_col', async () => {
+  it('Step 3~5: 生成计划→审阅→部署并验证 extra_col', async () => {
     await clickSchemaDiffNext();
     await clickSchemaDiffGeneratePlan();
     let body = await $('body').getText();
