+++ e2e/specs/settings.ts
@@ -2,6 +2,7 @@ import { expect, browser, $ } from '@wdio/globals';
 import {
   backFromSettingsInMainWindow,
   captureJourneyStep,
+  closeExtraWindows,
   openSettingsInMainWindow,
 } from '../helpers.js';
 import { t } from '../i18n.js';
@@ -377,6 +378,79 @@ describe('Settings (SS-001~SS-006)', () => {
     expect(saved).toBeDefined();
     expect(saved.command).toBe('/usr/bin/true');
     expect(saved.env?.TEST_ENV).toBe('e2e-value');
+  });
+
+  // ── Persistence: language / font-size / confirm-delete (from settings-persistence.ts) ──
+
+  it('TC-SET-008: 语言切换应更新 UI 文本', async () => {
+    const settings = await invokeBackend<Record<string, unknown>>('get_settings');
+    const originalLang = settings.language;
+
+    const newLang = originalLang === 'zh-CN' ? 'en' : 'zh-CN';
+    await invokeBackend('save_settings', { settings: { ...settings, language: newLang } });
+    await browser.refresh();
+    await browser.pause(2000);
+
+    const body = await $('body').getText();
+    if (newLang === 'en') {
+      expect(
+        body.includes('Settings') ||
+          body.includes('Connection') ||
+          body.includes('New Connection') ||
+          body.includes('Query'),
+      ).toBe(true);
+    } else {
+      expect(
+        body.includes('设置') ||
+          body.includes('连接') ||
+          body.includes('新建连接') ||
+          body.includes('查询'),
+      ).toBe(true);
+    }
+
+    // Restore
+    await invokeBackend('save_settings', { settings: { ...settings, language: originalLang } });
+    await browser.refresh();
+    await browser.pause(1500);
+  });
+
+  it('TC-SET-009: 编辑器字体大小设置应持久化', async () => {
+    const settings = await invokeBackend<Record<string, unknown>>('get_settings');
+    const originalFontSize = settings.editorFontSize;
+
+    const newFontSize = originalFontSize === 14 ? 16 : 14;
+    await invokeBackend('save_settings', {
+      settings: { ...settings, editorFontSize: newFontSize },
+    });
+    await browser.refresh();
+    await browser.pause(1500);
+
+    const persisted = await invokeBackend<Record<string, unknown>>('get_settings');
+    expect(persisted.editorFontSize).toBe(newFontSize);
+
+    // Restore
+    await invokeBackend('save_settings', {
+      settings: { ...settings, editorFontSize: originalFontSize },
+    });
+  });
+
+  it('TC-SET-010: 确认删除开关应持久化', async () => {
+    const settings = await invokeBackend<Record<string, unknown>>('get_settings');
+    const originalConfirm = settings.confirmOnDelete;
+
+    await invokeBackend('save_settings', {
+      settings: { ...settings, confirmOnDelete: !originalConfirm },
+    });
+    await browser.refresh();
+    await browser.pause(1500);
+
+    const persisted = await invokeBackend<Record<string, unknown>>('get_settings');
+    expect(persisted.confirmOnDelete).toBe(!originalConfirm);
+
+    // Restore
+    await invokeBackend('save_settings', {
+      settings: { ...settings, confirmOnDelete: originalConfirm },
+    });
   });
 
   // ── Restore defaults ──
