+++ e2e/wdio.conf.ts
@@ -34,9 +34,7 @@ export const config: WebdriverIO.Config = {
       './specs/schema-diff-window.ts',
       './specs/data-sync-window.ts',
       './specs/connection-edge-cases.ts',
-      './specs/settings-persistence.ts',
       './specs/window-operations.ts',
-      './specs/unified-main-window.ts',
       './specs/unified-tab-bar.ts',
       './specs/plugins.spec.ts',
     ],
@@ -53,8 +51,7 @@ export const config: WebdriverIO.Config = {
       './specs/object-browser.ts',
       './specs/data-types.ts',
       './specs/mysql.ts',
-      './specs/mysql-multi-database.ts',
-      './specs/postgres-multi-database.ts',
+      './specs/multi-database.ts',
       './specs/data-sync-real.ts',
       './specs/data-sync-edge-cases.ts',
       './specs/client-parity.ts',
@@ -62,7 +59,7 @@ export const config: WebdriverIO.Config = {
       './specs/sql-multi-tab.ts',
       './specs/schema-tree-completeness.ts',
       './specs/table-batch-ops.ts',
-      './specs/workflow-lifecycle.ts',
+      './specs/workflow.ts',
       './specs/er-diagram.ts',
       './specs/data-transfer-window.ts',
       './specs/data-transfer-type-mapping.ts',
